Use shfmt to format bash and bats scripts

+++ test.sh
@@ -1,12 +1,14 @@
 #!/usr/bin/env bash
 
 bats() {
-  docker run -it -v "${PWD}:/code" --rm bats/bats:latest "$@"
+  if (($#)); then
+    docker run -it -v "${PWD}:/code" --rm bats/bats:latest "$@"
+  fi
 }
 
 if (($#)); then
   bats "$@"
 else
   export -f bats
-  find test/ -type f -name '*.bats' -not -regex 'test/helpers/.*' -exec bash -c 'bats -j 4 --no-parallelize-within-files "$@"' {} +
+  find test/ -type f -name '*.bats' -not -regex 'test/helpers/mocks/.*' -exec bash -c 'bats -j 4 --no-parallelize-within-files "$@"' {} +
 fi
